Add support for Django 1.9
