+++ tox.ini
@@ -1,9 +1,10 @@
 [tox]
-envlist = py{27,34,35}-dj{18}
+envlist = py{27,34,35}-dj{18,19}
 
 [testenv]
 deps =
     dj18: django==1.8
+    dj19: django==1.9
     py27: mock
     factory_boy
     coverage
